feat: Refactor package structure and add User model for authentication
