+++ app/src/main/java/com/moviles/servitech/core/navigation/NavigationWrapper.kt
@@ -5,7 +5,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.moviles.servitech.view.LoginScreen
+import com.moviles.servitech.view.auth.LoginScreen
 
 @Composable
 fun NavigationWrapper() {
