+++ app/src/main/java/com/moviles/servitech/network/responses/LoginResponse.kt
@@ -1,6 +1,9 @@
 package com.moviles.servitech.network.responses
 
+import com.moviles.servitech.model.User
+
 data class LoginResponse(
+    val user: User,
     val token: String,
     val expiresIn: Int
 )
