--- app/src/main/java/com/moviles/servitech/network/utils/Result.kt
@@ -1,7 +1,0 @@
-package com.moviles.servitech.network.utils
-
-sealed class Result<out T> {
-    data class Success<out T>(val data: T) : Result<T>()
-    data class Error(val message: String, val code: Int) : Result<Nothing>()
-    object Loading : Result<Nothing>() // For loading state
-}
